fix type conversion error (on windows) unsigned/signed affecting timestamps on rrd_dump.
git-svn-id: svn://svn.oetiker.ch/rrdtool/trunk/program@1744 a5681a0c-68f1-0310-ab6d-d61299d08faa

+++ src/rrd_dump.c
@@ -394,7 +394,7 @@ static int rrd_dump_opt_r(
         rrd_seek(rrd_file, (rra_start + (rrd.rra_ptr[i].cur_row + 1)
                             * rrd.stat_head->ds_cnt
                             * sizeof(rrd_value_t)), SEEK_SET);
-        timer = -(rrd.rra_def[i].row_cnt - 1);
+        timer = -(long)(rrd.rra_def[i].row_cnt - 1);
         ii = rrd.rra_ptr[i].cur_row;
         for (ix = 0; ix < rrd.rra_def[i].row_cnt; ix++) {
             ii++;
